+++ environment.py
@@ -1,12 +1,12 @@
 import os
 
-from scripts.conf import BUILD_CLANG, BUILD_COMPILER_RT_MUSL, \
+from conf import BUILD_CLANG, BUILD_COMPILER_RT_MUSL, \
     BUILD_COMPILER_RT_NEWLIB, BUILD_LIBCXX, BUILD_LIBCXXABI, BUILD_LIBUNWIND, \
     BUILD_MUSL, BUILD_MUSL_HEADERS, BUILD_NEWLIB, BUILD_ROOT, MUSL_TRIPLE, \
     NEWLIB_TRIPLE, SRC_COMPILER_RT, SRC_LIBCXX, SRC_LIBCXXABI, SRC_LIBUNWIND, \
     SRC_LINUX, SRC_LLVM, SRC_MUSL, SRC_NEWLIB, SRC_ROOT, TARBALL_ROOT, \
     TARGETS_TO_BUILD, TOOLS_TO_BUILD
-from scripts.user_input import get_build_type, get_prefix
+from user_input import get_build_type, get_prefix
 
 
 def init_env():
