+++ preparation/download.py
@@ -2,9 +2,9 @@ import requests
 
 from pathlib import Path
 
-from scripts.conf import LINUX_TARBALL, LINUX_URL, MUSL_TARBALL, MUSL_URL, \
+from conf import LINUX_TARBALL, LINUX_URL, MUSL_TARBALL, MUSL_URL, \
     NEWLIB_TARBALL, NEWLIB_URL, TARBALL_ROOT
-from scripts.record_step import is_linux_headers_downloaded, \
+from utils.record_step import is_linux_headers_downloaded, \
     is_musl_downloaded, is_newlib_downloaded, linux_headers_downloaded, \
     musl_downloaded, musl_headers_downloaded, newlib_downloaded
