+++ preparation/patch.py
@@ -1,6 +1,6 @@
-from scripts.conf import PATCH_NEWLIB_C99, SRC_NEWLIB
-from scripts.record_step import is_newlib_patched, newlib_patched
-from scripts.run_shell import run_shell
+from conf import PATCH_NEWLIB_C99, SRC_NEWLIB
+from utils.record_step import is_newlib_patched, newlib_patched
+from utils.run_shell import run_shell
 
 
 def patch_src(src, patch):
@@ -20,8 +20,12 @@ def patch_newlib():
         newlib_patched()
 
 
+def patch_srouces():
+    patch_newlib()
+
+
 if __name__ == '__main__':
-    from scripts.environment import init_env
+    from environment import init_env
 
     init_env()
-    patch_newlib()
+    patch_srouces()
