+++ process/linux_headers.py
@@ -1,9 +1,9 @@
-from scripts.conf import SRC_LINUX
-from scripts.record_step import is_linux_headers_configured, \
+from conf import SRC_LINUX
+from user_input import get_libc_type
+from utils.record_step import is_linux_headers_configured, \
     is_linux_headers_installed, linux_headers_configured, \
     linux_headers_installed
-from scripts.run_shell import run_shell
-from scripts.user_input import get_libc_type
+from utils.run_shell import run_shell
 
 
 def do_config():
