+++ process/musl.py
@@ -1,8 +1,8 @@
-from scripts.conf import BUILD_MUSL
-from scripts.record_step import is_musl_built, is_musl_configured, \
+from conf import BUILD_MUSL
+from user_input import get_libc_type
+from utils.record_step import is_musl_built, is_musl_configured, \
     is_musl_installed, musl_built, musl_configured, musl_installed
-from scripts.run_shell import run_shell
-from scripts.user_input import get_libc_type
+from utils.run_shell import run_shell
 
 pre_config_args = \
     '''
