+++ process/musl_headers.py
@@ -1,9 +1,9 @@
-from scripts.conf import BUILD_MUSL_HEADERS
-from scripts.record_step import is_musl_headers_configured, \
+from conf import BUILD_MUSL_HEADERS
+from user_input import get_libc_type
+from utils.record_step import is_musl_headers_configured, \
     is_musl_headers_installed, musl_headers_configured, \
     musl_headers_installed
-from scripts.run_shell import run_shell
-from scripts.user_input import get_libc_type
+from utils.run_shell import run_shell
 
 config_args = \
     '''
